+++ Entity/JobTiming.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Dtc\QueueBundle\Entity;
+
+use Dtc\QueueBundle\Model\JobTiming as BaseJobTiming;
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\Entity
+ * @ORM\Table(name="dtc_queue_job_timing", indexes={@ORM\Index(name="job_timing_finished_at", columns={"finished_at"})})
+ */
+class JobTiming extends BaseJobTiming
+{
+    /**
+     * @ORM\Column(type="bigint")
+     * @ORM\Id
+     * @ORM\GeneratedValue(strategy="AUTO")
+     */
+    protected $id;
+
+    /**
+     * @ORM\Column(type="datetime")
+     */
+    protected $finishedAt;
+
+    /**
+     * @return mixed
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * @param mixed $id
+     */
+    public function setId($id)
+    {
+        $this->id = $id;
+    }
+}