+++ Model/JobTiming.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Dtc\QueueBundle\Model;
+
+class JobTiming
+{
+    protected $finishedAt;
+
+    /**
+     * @return mixed
+     */
+    public function getFinishedAt()
+    {
+        return $this->finishedAt;
+    }
+
+    /**
+     * @param mixed $finishedAt
+     */
+    public function setFinishedAt($finishedAt)
+    {
+        $this->finishedAt = $finishedAt;
+    }
+}