+++ Model/RunManager.php
@@ -4,16 +4,24 @@ namespace Dtc\QueueBundle\Model;
 
 class RunManager
 {
-    /** @var string|null */
+    /** @var string */
     protected $runClass;
 
-    public function __construct($runClass)
+    /** @var string */
+    protected $jobTimingClass;
+
+    /** @var bool */
+    protected $recordTimings;
+
+    public function __construct($runClass, $jobTimingClass, $recordTimings)
     {
         $this->runClass = $runClass;
+        $this->jobTimingClass = $jobTimingClass;
+        $this->recordTimings = $recordTimings;
     }
 
     /**
-     * @return null|string
+     * @return string
      */
     public function getRunClass()
     {
@@ -21,7 +29,7 @@ class RunManager
     }
 
     /**
-     * @param null|string $runClass
+     * @param string $runClass
      */
     public function setRunClass($runClass)
     {
@@ -36,5 +44,35 @@ class RunManager
     public function pruneArchivedRuns(\DateTime $olderThan)
     {
         throw new \Exception('not supported');
+    }
+
+    /**
+     * @param \DateTime $olderThan
+     *
+     * @return int Number of archived runs pruned
+     */
+    public function pruneJobTimings(\DateTime $olderThan)
+    {
+        throw new \Exception('not supported');
+    }
+
+    /**
+     * @returnnull|string
+     */
+    public function getJobTimingClass()
+    {
+        return $this->jobTimingClass;
+    }
+
+    /**
+     * @param string $jobTimingClass
+     */
+    public function setJobTimingClass($jobTimingClass)
+    {
+        $this->jobTimingClass = $jobTimingClass;
+    }
+
+    public function recordJobRun(Job $job)
+    {
     }
 }