+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### 2.3.0
+   * Added a new JobTiming entity/document to separately track timing information for each run
 ### 2.2.0
    * Create a RunManager for tracking Runs
    * Add a new config parameter run_manager so that a job queue can have a separate store-type for Runs